More support for the SERVICE environment.
git-svn-id: https://llvm.org/svn/llvm-project/vmkit/trunk@59517 91177308-0d34-0410-b5e6-96231b3b80d8

+++ include/mvm/VirtualMachine.h
@@ -46,7 +46,7 @@ public:
   
 
 #ifdef ISOLATE
-  uint32 IsolateID;
+  size_t IsolateID;
 #endif
 
 #ifdef SERVICE
